MA-20147 Make libcamera3 works ok on android-12
Purpose:
  Since it uses same HAL interfaces as libcamera's HAL, it can offert a good
  reference for bringing up libcamera's HAL on Android.

To pass build:
. Use libdmabufheap instead libion.
. Use fsl::Allocator instead of fsl::IonAllocator.
. Remove unsupported VIDIOC_DBG_G_CHIP_IDENT.

To run ok:
. Use 64bit for MessageQueue::arg0.

Note:
. Need use camera-provider-2.4, since 2.7 is totally from GCH(google camera HAL).
. Since it uses vidCap.driver as camera name, to run ok, in json config,
  need change "camera_name" from "mxc_isi.0.capture" to "mxc-isi-cap".

Reviewed-on: http://androidsource.ap.freescale.net/project/18460

+++ libcamera3/MessageQueue.h
@@ -62,10 +62,10 @@ public:
 class CMessage : public LightRefBase<CMessage>{
 public:
     int32_t what;
-    int32_t arg0;
+    uint64_t arg0;
 
     CMessage(int32_t what,
-             int32_t arg0 = 0)
+             uint64_t arg0 = 0)
         : what(what), arg0(arg0) {}
 
     virtual ~CMessage() {}
